Add files via upload
This is a document analyzer which Gives the Metadata about the Document and summarizes each document . it classifies each document based upon the context into categories like Resume , Article , invoice. This performs Abstractive Summarization using the transformers from the Hugging face. the model name is "sshleifer/distilbart-cnn-12-6" A lightweight but powerful Context summarizer.  
The model is trained on a lot of resumes , invoices and articles for 100% Accuracy. 
To run Refer the Readme file.